basic interpolation

+++ app/src/main/java/com/example/majo/merging/CartesianPoint.java
@@ -4,13 +4,13 @@ package com.example.majo.merging;
  * Created by majo on 14-Dec-14.
  */
 public class CartesianPoint {
-    public int x;
-    public int y;
+    public double x;
+    public double y;
 
     public CartesianPoint(){
 
     }
-    public CartesianPoint(int x, int y){
+    public CartesianPoint(double x, double y){
         this.x = x;
         this.y = y;
     }
